% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_mobster.R
\name{analyze_mobster}
\alias{analyze_mobster}
\title{Compute MOBSTER fits for the available samples.}
\usage{
analyze_mobster(x, samples = x$samples, ...)
}
\arguments{
\item{x}{A mvMOBSTER \code{mbst_data} object.}

\item{samples}{The samples on which MOBSTER will be run.}

\item{...}{Extra parameters that will be forwarded to
\code{mobster_fit} function from package \code{mobster}.}
}
\value{
A new object with MOBSTER fits for the required
samples in \code{x$fit.MOBSTER}.
}
\description{
This function wraps a call to the \code{mobster_fit} function
from package \code{mobster}, which implements the MOBSTER
statistical model. This is a combination of one Pareto power
law for a tumour tail, plus a mixture of Beta distributions.
The function applies the fitting function to the samples stored in the
dataset \code{x}, one by one, and the returned object contains the
MOBSTER fits which can be inspected with the functions available
in the \code{mobster} package.
}
\examples{
TODO
}
