% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_VIBER.R
\name{analyze_VIBER}
\alias{analyze_VIBER}
\title{Compute VIBER Binomial fits from counts data.}
\usage{
analyze_VIBER(x, samples = x$samples, ...)
}
\arguments{
\item{x}{A mvMOBSTER \code{mbst_data} object.}

\item{samples}{The samples on which VIBER will be run.}

\item{...}{Extra parameters that will be forwarded to
\code{variational_fit} function from package \code{VIBER}.}
}
\value{
A new object with a VIBER fit for the required
samples in \code{x$fit.Binomial}.
}
\description{
This function wraps a call to the \code{variational_fit} function
from package \code{VIBER}, which implements a variational Binmial
mixture model to cluster read counts. See the package manual for
\code{VIBER} to see the parameters that can be forwarded, and the
plots that can be generated afterwards.
}
\examples{
TODO
}
