% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_data.R
\name{Annotations}
\alias{Annotations}
\title{Extract annotation values.}
\usage{
Annotations(x, ids = keys(x))
}
\arguments{
\item{x}{A mvMOBSTER \code{mbst_data} object.}

\item{ids}{The IDs of the mutations to extract, all by default.}
}
\value{
A tibble of the required entries
}
\description{
Extract from the internal representation of an object all the entries
that refer to the annotation values. These are all values which are not
any of VAF, DP or NV, or any of the location coordinates for the input
mutations. As for the other getters, you can subset by sample and mutation
ID; by default all entries are returbed. The output is a tibble; no other
transformations are executed.
}
\examples{
TODO
}
