% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_getters.R
\name{BClusters}
\alias{BClusters}
\title{Getter for VIBER Binomial clusters}
\usage{
BClusters(x, annotations = FALSE)
}
\arguments{
\item{x}{A mvMOBSTER \code{mbst_data} object.}

\item{annotations}{If the annotations available for the mutations should be returned.}
}
\value{
VIBER clustering assigments for every mutation, if available.
}
\description{
Getter for VIBER Binomial clusters
}
\examples{
TODO
}
