% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chr_cytoband_hg19}
\alias{chr_cytoband_hg19}
\title{Cytobands for hg19 chromosome}
\format{A tibble that represents the cytobands for hg19 chromosome
(chr, from, to, cytoband and region).}
\usage{
data(chr_cytoband_hg19)
}
\description{
Cytobands for hg19 chromosome
}
\examples{
data(chr_cytoband_hg19)
head(chr_cytoband_hg19)

}
\references{
hg19
}
\keyword{datasets}
