% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_getters.R
\name{Clusters}
\alias{Clusters}
\title{Getters for clustering results.}
\usage{
Clusters(x, cluster, annotations = FALSE)
}
\arguments{
\item{x}{A mvMOBSTER \code{mbst_data} object.}

\item{cluster}{Use one of MOBSTER (\code{M}), sciClone (\code{S}), pyClone (\code{P})
and Binomial (\code{B}).}

\item{annotations}{If the annotations available for the mutations should be returned.}
}
\value{
Different types of clustering assigments for every mutation.
}
\description{
Getters for clustering results.
}
\examples{
TODO
}
