% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_wide_format.R
\name{Data_table}
\alias{Data_table}
\title{Extract all the data stored in an \code{evoverse} dataset, in the wide format.}
\usage{
Data_table(x, ids = keys(x), samples = x$samples, coordinates = TRUE)
}
\arguments{
\item{x}{An `evoverse` object.}

\item{ids}{The id of the mutations to extract, all by default.}

\item{samples}{The samples for which we extract the data, all by default.}

\item{coordinates}{If `TRUE`, returns also the genomic coordinates of the annotated mutations (`TRUE` by default).}
}
\value{
A spread tibble of the required entries.
}
\description{
This is just a wrapper to a combined call of function \link{\code{VAF_table}}, \link{\code{DP_table}} and
\link{\code{NV_table}}. The resulting outputs are joind by id; usual subsetting options are available.
The output contains also the annotations from each available mutation, obtained with \link{\code{Annotations}}.
}
\examples{
data('example_evoverse')
Data_table(example_evoverse)
}
