% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Purity.R
\name{estimate_purity}
\alias{estimate_purity}
\title{Estimate the purity of}
\usage{
estimate_purity(x, peak.range = c(0.2, 0.5), m.peak = 10)
}
\arguments{
\item{x}{A mvMOBSTER \code{mbst_data} object.}

\item{peak.range}{A range of VAF values where
the peaks are searched for.}

\item{m.peak}{The neighbour which are compared
against when searching for a peak.}
}
\value{
A \code{tibble} with the peaks per sample
}
\description{
From the peaks of the VAF distributions
for each sample in \code{x}, this function
estimates the purity looking for the
highest peak in a range \code{peak.range}
which is by default (0.2, 0.5, i.e., from 40%
to 100% tumour content). An extra parameter
\code{m.peak} defines the contour of each
peak; change this parameter to get different
estimates.
}
\examples{
TODO
}
