% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evo_pipe.R
\name{evo_pipe}
\alias{evo_pipe}
\title{Data analysis template}
\usage{
evo_pipe(x, fun_data_load, fun_compute, fun_postprocess, cache = NULL)
}
\arguments{
\item{x}{A sample character id.}

\item{fun_data_load}{A function applied to sample to obtain an input object;
for instance loads a CSV from a folder indexed by the sample id.}

\item{fun_compute}{A function that implements the actual computation on
to the input object; for instance it clusters the input. All parameters must
be made explicit inside this function. .}

\item{fun_postprocess}{A function to compute the final result from the
analysis output, and its input.}

\item{cache}{By default \code{NULL}, which turns off the cache. If this is a
filename, and the file exists, no computation is skip.}
}
\value{
A list with the sample and results from each each step of the
pipeline
}
\description{
A three-steps higher-order pipeline; loads input data, computes a result and
postprocess it. It implements an cache system to avoid computing when results
are available (optional).

In tidy language the pipeline is roughly this, plus the caching system

fun_data_load %>%
   fun_compute %>%
   fun_postprocess
}
