% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{evoparse_mutect_mutations}
\alias{evoparse_mutect_mutations}
\title{Pares Platypus mutation data.}
\usage{
evoparse_mutect_mutations(x, sample = NULL)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Pares Platypus mutation data.
}
