% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{evoparse_platypus_mutations}
\alias{evoparse_platypus_mutations}
\title{Pares Platypus mutation data.}
\usage{
evoparse_platypus_mutations(x)
}
\arguments{
\item{x}{}
}
\description{
Pares Platypus mutation data.
}
