% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{evoparse_Sequenza_CNAs}
\alias{evoparse_Sequenza_CNAs}
\title{Parese Sequenza Copy Number data.}
\usage{
evoparse_Sequenza_CNAs(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Parese Sequenza Copy Number data.
}
