% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_evoverse}
\alias{example_evoverse}
\title{Example \code{evoverse} dataset.}
\format{A simple dataset created and analyzed with the
\code{evoverse} package.}
\usage{
data(example_evoverse)
}
\description{
Example \code{evoverse} dataset.
}
\examples{
data(example_evoverse)
head(example_evoverse)

}
\keyword{datasets}
