% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_input_formats}
\alias{example_input_formats}
\title{Example data in the wide and long formts supported by mvMOBSTER}
\format{4 tibbles, wide/long and mutations/copy number that
show the format required to create a mvMOBSTER object with
function \link{\code{dataset}}.}
\usage{
data(example_input_formats)
}
\description{
Example data in the wide and long formts supported by mvMOBSTER
}
\examples{
data(example_input_formats)
head(example_input_formats)
}
\keyword{#'}
\keyword{datasets}
