% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_mvmobster}
\alias{example_mvmobster}
\title{Example data in mvMOBSTER format.}
\format{A tibble that represents the cytobands for hg19 chromosome
(chr, from, to, cytoband and region).}
\usage{
data(example_mvmobster)
}
\description{
Example data in mvMOBSTER format.
}
\examples{
data(example_mvmobster)
head(example_mvmobster)

}
\references{
hg19
}
\keyword{datasets}
