% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_downsample.R
\name{filter_downsample}
\alias{filter_downsample}
\title{Downample data.}
\usage{
filter_downsample(x, n)
}
\arguments{
\item{x}{A dataset of class \code{"mbst_data"}.}

\item{N}{The number of mutations to retain. If \code{x} contains less
than \code{N} mutations, \code{x} is returned.}
}
\value{
A dataset with at most \code{N} mutations.
}
\description{
Randomly downsample the data contained in a dataset to a
certain number of mutations.
}
\examples{
TODO
}
