% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dp_range.R
\name{filter_dp_range}
\alias{filter_dp_range}
\title{Filter mutations outside a range of coverage values.}
\usage{
filter_dp_range(x, min.DP, max.DP)
}
\arguments{
\item{x}{An `evoverse` object.}

\item{min.DP}{Minimum DP}

\item{max.DP}{Maximum DP}
}
\value{
A dataset with no mutations outside the
required DP range.
}
\description{
@description This function filters mutations by DP (depth).
}
\details{
The filter deletes all mutations that, at least in one sample,
have read depth DP outside a desired range.
}
\examples{
data('example_evoverse')

N(example_evoverse)

N(filter_dp_range(example_evoverse, 50, 200))
}
