% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_filters.R
\name{filter_dprange}
\alias{filter_dprange}
\title{Filter all mutations in a DP range}
\usage{
filter_dprange(x, min.DP, max.DP)
}
\arguments{
\item{x}{A dataset of class \code{"mbst_data"}.}

\item{min.DP}{Minimum DP}

\item{max.DP}{Maximum DP}
}
\value{
A dataset with no mutations outside the
required VAF range.
}
\description{
Filter all mutations that, when they have
DP above 0, they are not in a certain range, given by
\code{min.DP}, \code{max.DP}.
}
\examples{
TODO
}
