% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_filters.R
\name{filter_minnv}
\alias{filter_minnv}
\title{Set to 0 NV entries below a cutoff.}
\usage{
filter_minnv(x, cutoff)
}
\arguments{
\item{x}{A dataset of class \code{"mbst_data"}.}

\item{cutoff}{The cutoff to set to 0 VAF/NV entries.}
}
\value{
A dataset without mutations with NV below \code{cutoff}.
}
\description{
@description Find mutations with NV below \code{cutoff}, set to 0 both
their VAF and NV entries and retain the coverage at the locus (DP).
}
\examples{
TODO
}
