% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_nv_range.R
\name{filter_nv_range}
\alias{filter_nv_range}
\title{Filter mutations outside a range of NV values.}
\usage{
filter_nv_range(x, min.NV, max.NV)
}
\arguments{
\item{x}{An `evoverse` object.}

\item{min.NV}{Minimum NV}

\item{max.NV}{Maximum NV}
}
\value{
A dataset with no mutations outside the
required NV range.
}
\description{
@description This function filters mutations by NV (number of variants).
}
\details{
The filter deletes all mutations that, at least in one sample,
a number of variants NV outside a desired range.
}
\examples{
data('example_evoverse')

N(example_evoverse)

N(filter_nv_range(example_evoverse, 50, 100))
}
