% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_tails.R
\name{filter_tails}
\alias{filter_tails}
\title{Compute projected mutations from MOBSTER fits.}
\usage{
filter_tails(x)
}
\arguments{
\item{x}{A dataset of class \code{"mbst_data"}.}
}
\value{
A dataset with projected mutations from MOBSTER fits.
}
\description{
After MOBSTER has been run on each sample stored in
\code{x}, the fits can be used to projected mutations that are
found to be part of a tail at least onece. For these mutations
their VAF and NV levels are forced to 0 in all samples, and
the mutations are then removed.
}
\examples{
TODO
}
