% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_vaf_nv_min.R
\name{filter_vaf_nv_min}
\alias{filter_vaf_nv_min}
\title{Impose minimum VAF and NV (read counts) value for mutations.}
\usage{
filter_vaf_nv_min(x, min.vaf, min.nv)
}
\arguments{
\item{x}{An `evoverse` object.}

\item{min.vaf}{Minimum VAF.}

\item{min.NV}{Minimum NV}
}
\value{
A dataset where any VAF entry below \code{min.vaf}, and any NV entry below
\code{min.nv} are changed to 0.
}
\description{
This function sets the VAF and NV of mutations below a minimum cutoff to 0.
The mutation is not removed from the dataset.
}
\examples{
data('example_evoverse')

N(example_evoverse)

N(filter_vaf_nv_min(example_evoverse, .05, 5))

Data_table(example_evoverse)

Data_table(filter_vaf_nv_min(example_evoverse, .05, 5))
}
