% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_vaf_range.R
\name{filter_vaf_range}
\alias{filter_vaf_range}
\title{Filter mutations with VAF outside a range.}
\usage{
filter_vaf_range(x, min.vaf, max.vaf)
}
\arguments{
\item{x}{An `evoverse` object.}

\item{min.vaf}{Minimum VAF.}

\item{max.vaf}{Maximum VAF.}
}
\value{
A dataset with no mutations outside the
required VAF range.
}
\description{
This function filters mutations by VAF.

The filter deletes all mutations that, when they have
VAF above 0 in a sample (and so they are "called" in that sample),
have a VAF value outside a desired range.
}
\examples{
data('example_evoverse')

N(example_evoverse)

N(filter_vafrange(example_evoverse, .1, .6))
}
