% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_long_format.R
\name{keys}
\alias{keys}
\title{Get the keys used to index the mutations in the data.}
\usage{
keys(x)
}
\arguments{
\item{x}{An `evoverse` object.}
}
\value{
The keys that index stored mutations.
}
\description{
Every mutation has its own key which is used to
access and track the mutation inside an  \code{evoverse} object.

With this function you can extract the keys:  many functions accept
them as parameter to extract data.
}
\examples{
data('example_evoverse')

my_keys = keys(example_evoverse)

# Example usage
VAF(example_evoverse, ids = my_keys[1])
}
