% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_getters.R
\name{MClusters}
\alias{MClusters}
\title{Gettes for MOBSTER clusters}
\usage{
MClusters(x, annotations = FALSE)
}
\arguments{
\item{x}{A mvMOBSTER \code{mbst_data} object.}

\item{annotations}{If the annotations available for the mutations should be returned.}
}
\value{
MOBSTER clustering assigments for every mutation, if available.
}
\description{
Gettes for MOBSTER clusters
}
\examples{
TODO
}
