% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{mobster_dataset}
\alias{mobster_dataset}
\title{Create a dataset.}
\usage{
mobster_dataset(mutations, segments, samples, purity,
  description = "My multi-region MOBSTER dataset", N.min = 500,
  offset_around_centromers = 1e+06)
}
\arguments{
\item{mutations}{Dataframe of mutations. Must contain mutation
coordinate as chr, from and to. Plus, for every sample named
\code{x}, it must contain columns \code{x.DP}, \code{x.NV}
and \code{x.VAF} with the values of tumour depth (DP), number
of reads with the mutant allele (NV) and variant allele
frequency (VAF). Exra columns are retained.}

\item{segments}{Copy number segments, must contain segment
coordinate as chr, from and to. for every sample named
\code{x}, it must contain columns \code{x.minor} and
\code{x.Major} with the minor and major number of copies
of the segment.}

\item{samples}{Vector of samples names.}

\item{purity}{Vector of purities, with samples as names.}

\item{description}{Dataset synopsis.}

\item{N.min}{Minimum number of mutations that need to map to
a segment. If a segment has less than \code{N.min}, the
segment is rejected.}

\item{offset_around_centromers}{When mapping mutations to
around segments, exclude this offset around the centromers
of each chromosome (reference coordinates hg19).}
}
\value{
An object that represents a dataset from class
\code{mbs_data}.
}
\description{
Create a mvMOBSTER dataset that can contain
mutation data for a one, or more, bulk samples of the same
tumour. The method requires mutations, copy number segments
and sample purity to adjust the raw VAF for copy state and
tumour content. The adjusted VAF of clonal mutations in
diploid region is expect to be around 0.5 (heterozygous
mutation) that corresponds to 100% of cancer cells fraction.
Basic filtering options are provided.
}
\examples{
data(example_mvmobster)
TODO
}
