% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_wrappers.R
\name{MOBSTER_fit_pyClone}
\alias{MOBSTER_fit_pyClone}
\title{Compute pyClone fits for the available samples.}
\usage{
MOBSTER_fit_pyClone(x, iterations = 10000, burnin = 1000, seed = 100)
}
\arguments{
\item{x}{A mvMOBSTER \code{mbst_data} object.}

\item{iterations}{Number of pyClone iterations (MCMC).}

\item{burnin}{Burnin for pyClone MCMC.}

\item{seed}{Random seed.}
}
\value{
A new object with pyClone fits in \code{x$fit.pyClone}.
}
\description{
This function wraps a call to pyClone, a python
tool which must be installed and accessyble through a
system call within R. The function applies the fitting
function to the samples stored in the dataset \code{x}, and the
returned object is built scanning pyClone outputs and loading
the relevant information.
}
\examples{
TODO
}
