% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_wrappers.R
\name{mobster_fit_sciClone}
\alias{mobster_fit_sciClone}
\title{Compute sciClone fits for the available samples.}
\usage{
mobster_fit_sciClone(x, minimumDepth = 0, maximumClusters = 2 *
  length(x$samples), ...)
}
\arguments{
\item{x}{A mvMOBSTER \code{mbst_data} object.}

\item{minimumDepth}{Mutations below this threshold will be removed.}

\item{maximumClusters}{Maximum number of clusters to return.}

\item{...}{Extra parameters that will be forwarded to
\code{mobster_fit} function from package \code{mobster}.}
}
\value{
A new object with sciClone fits in \code{x$fit.sciClone}.
}
\description{
This function wraps a call to the sciClone main fitting function
from package \code{sciClone}. The function applies the fitting
function to the samples stored in the dataset \code{x}, and the
returned object contains can be inspected
with the functions available in the \code{sciClone} package.
}
\examples{
TODO
}
