% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_filters.R
\name{mobster_flt_vafrange}
\alias{mobster_flt_vafrange}
\title{Filter all mutations in a VAF range}
\usage{
mobster_flt_vafrange(x, min.vaf, max.vaf)
}
\arguments{
\item{x}{A dataset of class \code{"mbst_data"}.}

\item{min.vaf}{Minimum VAF.}

\item{max.vaf}{Maximum VAF}
}
\value{
A dataset with no mutations outside the
required VAF range.
}
\description{
Filter all mutations that, when they have
VAF above 0, they are not in a certain range, given by
\code{min.vaf}, \code{max.vaf}.
}
\examples{
TODO
}
