% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_data_size.R
\name{N}
\alias{N}
\title{Return the number of mutations in the dataset.}
\usage{
N(x)
}
\arguments{
\item{x}{An `evoverse` object.}
}
\value{
The number of mutations in the dataset.
}
\description{
Return the number of mutations in the dataset.
}
\examples{
data('example_evoverse')
N(example_evoverse)
}
