% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_wide_format.R
\name{NV_table}
\alias{NV_table}
\title{Extract NV values in wide format (multi-column).}
\usage{
NV_table(x, ids = keys(x), samples = x$samples, suffix = ".NV")
}
\arguments{
\item{x}{An `evoverse` object.}

\item{ids}{The id of the mutations to extract, all by default.}

\item{samples}{The samples for which we extract the data, all by default.}

\item{suffix}{Every column will be called as \code{sample_name.VAF} when \code{suffix = '.NV'}.}
}
\value{
A spread tibble of the required entries.
}
\description{
This function spreads the result of the call to the \link{\code{NV}}
getter following the principles of wide data representation.
}
\examples{
data('example_evoverse')
NV_table(example_evoverse)
}
