% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_getters.R
\name{PClusters}
\alias{PClusters}
\title{Gettes for pyClone clusters}
\usage{
PClusters(x, annotations = FALSE)
}
\arguments{
\item{x}{A mvMOBSTER \code{mbst_data} object.}

\item{annotations}{If the annotations available for the mutations should be returned.}
}
\value{
pyClone clustering assigments for every mutation, if available.
}
\description{
Gettes for pyClone clusters
}
\examples{
TODO
}
