% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_clonal_mutations.R
\name{pipeline_clonal_mutations_purity}
\alias{pipeline_clonal_mutations_purity}
\title{Determine clonal mutations and sample purity from mutation data.}
\usage{
pipeline_clonal_mutations_purity(x, cutoff_lv = 0.7, N_min = 20, ...)
}
\arguments{
\item{x}{A dataframe of somatic mutations that can be processed with
MOBSTER, and that has the column \code{"chr"}.}

\item{cutoff_lv}{A desired starting cutoff to compute clustering
assignments with the latent variables probabilities above this threshold.
This value gets decreased progressively if not enouth mutations are
assignable to the cluster.}

\item{N_min}{Minimum number of mutations that we wish to have in a cluster
to consider it for this analysis.}

\item{...}{Parameters for the model fit that are forwarded to \code{\link{mobster_fit}}
from package [MOBSTER](https://caravagn.github.io/mobster/)}
}
\value{
A list of named objects. These contain the best fit to the data,
a summary table with the results of the heuristics and the best choice
for the clonal cluster, high-confidence clonal mutations and tumour purity.
}
\description{
This function implements a pipeline that uses MOBSTER to determine
a pool of clonal mutations for a tumour biopsy, and the purity of
the associated bulk sample.

This function does not work with Copy Number data, and uses only
the oberved raw VAF from a set of available somatic mutations. To use
also CNA data, please refer to \code{\link{}}

The pipeline implements two heuristics to improve the quality of the
output results:

\itemise{
\item determine a dynanmic cutoff of the latent variables for
MOBSTER hard clustering assignments, so to ensure that a minimum number
of mutations is assigned to the clonal cluster. This means that the
pipeline starts with an initial value for the cutoffs, which gets
decreased until the minimum number of mutations is reached. If the
mutational burden was insufficient, the best possible result is still returned.

\item determine if most of the mutations available in one cluster are
mapped to a subset of chromosome locations (i.e., they are clustered). This
suggests which clusters might be originated from Copy Number events (e.g. LOH)
that are not available in the input data. In practice, it checks if more than
60% of the mutations in one cluster map to less than 20% of the chromosomes
that have been used for this analysis (all the genome).
}
}
\examples{
Take one dataset with genomic locations
data('PD4120a_breast_sample', package = 'mobster')

# Tune the pipeline (auto_setup = 'FAST')
x = pipeline_clonal_mutations_purity(
 PD4120a_breast_sample$best$data,
 cutoff_lv = .7,
 auto_setup = 'FAST' # These go to MOBSTER mobster_fit
 )

print(x)
}
