% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIPELINE_QC_CALLS.R,
%   R/pipeline_qc_copynumbercalls.R
\name{pipeline_qc_copynumbercalls}
\alias{pipeline_qc_copynumbercalls}
\title{Pipeline for QC of somatic calls (copy number and mutations).}
\usage{
pipeline_qc_copynumbercalls(mutations, cna, purity, reference = "GRCh38",
  sample = "PAT00001", output = paste0(sample,
  "_evoverse_qcreport.pdf"), matching_epsilon_peaks = 0.025,
  smooth = FALSE, cex = 0.7, collate = FALSE)

pipeline_qc_copynumbercalls(mutations, cna, purity, reference = "GRCh38",
  sample = "PAT00001", output = paste0(sample,
  "_evoverse_qcreport.pdf"), matching_epsilon_peaks = 0.025,
  smooth = FALSE, cex = 0.7, collate = FALSE)
}
\arguments{
\item{mutations}{Somatic mutation calls, in \code{CNAqc} input format.}

\item{cna}{Absolute Copy Number segments, in \code{CNAqc} input format.}

\item{purity}{Tumour purity, in \code{CNAqc} input format.}

\item{reference}{A reference supported by package \code{CNAqc}.}

\item{sample}{A sample identifier; will appear in the plots.}

\item{matching_epsilon_peaks}{Epsilon to match peaks with \code{CNAqc}.}

\item{smooth}{If \code{TRUE}, smooths copy number data with \code{CNAqc}.}

\item{collate}{}
}
\description{
This is ...

All input data must be in the format supported by package \code{CNAqc}.
}
