% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIPELINE_DECONV_VAF_HIERARCHICAL.R
\name{pipeline_subclonal_deconvolution_hierarchical_VAF}
\alias{pipeline_subclonal_deconvolution_hierarchical_VAF}
\title{Title}
\usage{
pipeline_subclonal_deconvolution_hierarchical_VAF(
  x,
  karyotypes = c("1:0", "1:1", "2:0", "2:1", "2:2"),
  description = "VAF by karyotype hierarchical deconvolution sample",
  min_muts = 150,
  min_vaf = 0,
  enforce_QC_PASS = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{

}
