% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIPELINE_DECONV_VAF_KARYOTYPE.R
\name{pipeline_subclonal_deconvolution_VAF_karyotype}
\alias{pipeline_subclonal_deconvolution_VAF_karyotype}
\title{Pipeline to perform subclonal deconvolution with MOBSTER.}
\usage{
pipeline_subclonal_deconvolution_VAF_karyotype(
  x,
  karyotypes = c("1:0", "1:1", "2:0", "2:1", "2:2"),
  description = "VAF by karyotype deconvolution sample",
  min_muts = 150,
  N_max = 15000,
  enforce_QC_PASS = TRUE,
  ...
)
}
\arguments{
\item{x}{An object returned from function \code{pipeline_qc_copynumbercalls},
which refers to the QC analysis of somatic mutations, CNAs and tumour purity.}

\item{karyotypes}{Vector of ..}

\item{description}{A string description of this pipeline run, it will appear
in some of the produced plots.}

\item{min_muts}{Skip analysing karyotypes with less than \code{min_muts} mutations.}

\item{enforce_QC_PASS}{If \code{TRUE}, the pipeline will only analyze samples
that have PASS status from the data analysis QC}

\item{...}{Parameters passed \code{mobster_fit} in package \code{mobster}.}
}
\value{

}
\description{
For a list of input mutations, CNA and purity data from a tumour sample,
MOBSTER fits are computed to determine proportions of mutations in each
aneuploidy state (copy state). Assignments and plots are returned, and
an automatic QC pipeline (supervised logistic classifier) is used to
flag which fits pass quality check. The trained set for the classifier
consists in n = 377 colorectal whole genome samples.

Run \code{vignette("a3_Deconv_vaf_karyo", package = "evoverse")} to see
the corresponding vignette.
}
\examples{
# We use data released with the CNAqc package

print(x)

# We use x to run the pipeline
y = pipeline_subclonal_deconvolution_VAF_karyotype(x, auto_setup = 'FAST', N_max = 500)

print(y)
}
