% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIPELINE_DECONV_CCF.R
\name{plot.evopipe_ccf}
\alias{plot.evopipe_ccf}
\title{S3 plot for the deconvolution pipeline from raw VAF with karyotypes}
\usage{
\method{plot}{evopipe_ccf}(x, ...)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
S3 plot for the deconvolution pipeline from raw VAF with karyotypes
}
