% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIPELINE_DECONV_VAF_KARYOTYPE.R
\name{plot.evopipe_rawk}
\alias{plot.evopipe_rawk}
\title{S3 plot for the deconvolution pipeline from raw VAF with karyotypes}
\usage{
\method{plot}{evopipe_rawk}(x, ...)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
S3 plot for the deconvolution pipeline from raw VAF with karyotypes
}
