% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_plot.R
\name{plot.mbst_data}
\alias{plot.mbst_data}
\title{Plot a dataset (complex figure).}
\usage{
plot.mbst_data(x, scales = "fixed", cex = 1, ...)
}
\arguments{
\item{x}{A dataset of class \code{"mbst_data"}.}

\item{scales}{Scales for ggplot facets, default is \code{'fixed'} but \code{'free'} can also help.}

\item{cex}{Cex of the plots.}

\item{...}{}
}
\value{
A ggpubr figure object.
}
\description{
Create a complex figure via \code{ggpubr} assembling several plots. The figure
reports 1) the adjusted VAF (full spectrum), 2) the zoomed version of the adjusted
VAF for values below 0.20 (low-frequency spectrum), 3) the depth against the VAF
of each mutation and 4) the depth distribution.
}
\examples{
TODO
}
