% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_1D_VAF.R
\name{plot_1D_VAF}
\alias{plot_1D_VAF}
\title{Plot a 1-dimensional VAF histogram.}
\usage{
plot_1D_VAF(x, s = x$samples[1])
}
\arguments{
\item{x}{A `mvMOSTER` object.}

\item{s}{The sample name, by default `x$samples[1]` the first sample in the data.}
}
\description{
For an input sample, this plots a one dimensional
histogam of the input VAF, coloured by mutation karyotype.
}
