% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2D_VAF.R
\name{plot_2D_VAF}
\alias{plot_2D_VAF}
\title{Plot a 2-dimensional VAF scatterplot}
\usage{
plot_2D_VAF(x, s1 = x$samples[1], s2 = x$samples[2])
}
\arguments{
\item{x}{A `mvMOSTER` object.}

\item{s1}{The first sample name, by default `x$samples[1]` the first sample in the data.}

\item{s1}{The second sample name, by default `x$samples[1]` the second sample in the data.}
}
\description{
For 2 input samples, this plots a 2D scatterplot
of the input VAF in `s1` versus `s2`.
}
