% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2D_VAF_MOBSTER.R
\name{plot_2D_VAF_MOBSTER}
\alias{plot_2D_VAF_MOBSTER}
\title{Plot a 2-dimensional VAF scatterplot, augmented with MOBSTER cluster}
\usage{
plot_2D_VAF_MOBSTER(x, s1 = x$samples[1], s2 = x$samples[2],
  N = 10000, ...)
}
\arguments{
\item{x}{A `mvMOSTER` object.}

\item{s1}{The first sample name, by default `x$samples[1]` the first sample in the data.}

\item{N}{Maximum number of points to plot, the overall percentage is reported.}

\item{s1}{The second sample name, by default `x$samples[1]` the second sample in the data.}
}
\description{
For 2 input samples, this plots a 2D scatterplot
of the input VAF in `s1` versus `s2`. The colour of each point
represents the result of running MOBSTER on the input data.
}
