% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_extra.R
\name{plot_CNsegments}
\alias{plot_CNsegments}
\title{Plot copy number segments.}
\usage{
plot_CNsegments(x, samples = x$samples, ...)
}
\arguments{
\item{x}{An object of class \code{mbst_data}.}

\item{samples}{Required samples.}

\item{...}{Extra parameters passed to the internal
hidden function \code{plt_CNsegments}.}
}
\value{
A list of \code{ggarrange} objects named after samples.
}
\description{
Plots the copy number segments available for a required
set of samples. Returns a list of plots named after each
input sample via \code{samples}. For every sample it plots
the minor and major segments, highlights diploid regions
(kariotype minor = 1, Major = 1) and annotates on top the
number, the VAF and the DP of each annotated mutation. Each
returned object is a figure arranged via \code{ggarrange}.
}
\examples{
TODO
}
