% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_extra.R
\name{plot_count_kariotypes}
\alias{plot_count_kariotypes}
\title{Plot the counts of each kariotype.}
\usage{
plot_count_kariotypes(x, cex = 1)
}
\arguments{
\item{x}{An object of class \code{mbst_data}.}

\item{cex}{Cex of the plot.}
}
\value{
A ggplot object.
}
\description{
Plot the number of segments per sample and
per kariotype, in \code{x}, as a barplot.
}
\examples{
data("example_mvmobster")
}
