% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_icon_CNA.R
\name{plot_icon_CNA}
\alias{plot_icon_CNA}
\title{Title}
\usage{
plot_icon_CNA(
  x,
  colors = c(`0` = "steelblue", `1` = "forestgreen", `2` = "darkorange", `3` = "red",
    `3+` = "darkred")
)
}
\arguments{
\item{colors}{}
}
\value{

}
\description{
Title
}
