% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample.R
\name{plot_sample}
\alias{plot_sample}
\title{Plot a report of the data available for one sample}
\usage{
plot_sample(x, sample, N_CNA = 10000)
}
\arguments{
\item{x}{A mvMOBSTER object.}

\item{sample}{The sample to plot.}

\item{N_CNA}{The number of mutations to plot in the CNA segment
panel. This parameter is passed to the plotting functions of
the \code{CNAqc} package.}
}
\value{
A figure assembled from several \code{ggplot} objects.
}
\description{
For an input sample, plot a full page figure with
the histograms of VAF, DP and NV data. The joint scatter of
DP against VAF, as well as the CNA segments generated from the
\code{CNAqc} package.
}
\examples{
x = 3
}
