% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_extra.R
\name{plot_VAF_kariotypes}
\alias{plot_VAF_kariotypes}
\title{Plot the VAF per kariotype.}
\usage{
plot_VAF_kariotypes(x, range = c(0, Inf), cex = 1)
}
\arguments{
\item{x}{An object of class \code{mbst_data}.}

\item{range}{The VAF range (left to right), whichi is by default [0;+inf).}

\item{cex}{Cex of the plot.}
}
\value{
A ggplot object.
}
\description{
Plot the adjusted VAF histogram from the mutations
in \code{x}, split by kariotype.
}
\examples{
data("example_mvmobster")
}
