% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIPELINE_CHR_TIMING.R
\name{print.evopipe_ctime}
\alias{print.evopipe_ctime}
\title{S3 print for the chromosome timing pipeline results.}
\usage{
\method{print}{evopipe_ctime}(x, ...)
}
\arguments{
\item{x}{}
}
\description{
S3 print for the chromosome timing pipeline results.
}
