% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_print.R
\name{print.mbst_data}
\alias{print.mbst_data}
\title{Print for an \code{evoverse} dataset.}
\usage{
\method{print}{mbst_data}(x, ...)
}
\arguments{
\item{x}{An  \code{evoverse} object.}

\item{...}{S3 parameters}
}
\value{
Nothing.
}
\description{
The print reports basic statistics on the stored
fits available in the input object.
}
\examples{
data('example_evoverse')
print(example_evoverse)
}
