% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qc_deconvolution_monoclonal}
\alias{qc_deconvolution_monoclonal}
\title{QC pre-trained Logistic classifier for deconvolution with a single Beta component with MOBSTER
(monoclonal case).}
\format{A logistic classifier.... describe training ......}
\usage{
data(qc_deconvolution_monoclonal)
}
\description{
QC pre-trained Logistic classifier for deconvolution with a single Beta component with MOBSTER
(monoclonal case).
}
\examples{
print(qc_deconvolution_monoclonal)
}
\keyword{datasets}
