% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qc_deconvolution_polyclonal}
\alias{qc_deconvolution_polyclonal}
\title{QC pre-trained Logistic classifier for deconvolution with multiple Beta component with MOBSTER
(polyclonal case).}
\format{A logistic classifier.... describe training ......}
\usage{
data(qc_deconvolution_polyclonal)
}
\description{
QC pre-trained Logistic classifier for deconvolution with multiple Beta component with MOBSTER
(polyclonal case).
}
\examples{
print(qc_deconvolution_polyclonal)
}
\keyword{datasets}
