% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qc_timing_model}
\alias{qc_timing_model}
\title{QC pre-trained Logistic classifier for timing of CNA with MOBSTER.}
\format{A logistic classifier.... describe training ......}
\usage{
data(qc_timing_model)
}
\description{
QC pre-trained Logistic classifier for timing of CNA with MOBSTER.
}
\examples{
print(qc_timing_model)
}
\keyword{datasets}
