% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_data.R
\name{S}
\alias{S}
\title{Return the number of samples in the dataset.}
\usage{
S(x)
}
\arguments{
\item{x}{A mvMOBSTER \code{mbst_data} object.}
}
\value{
The number of samples in the dataset.
}
\description{
Return the number of samples in the dataset.
}
\examples{
TODO
}
