% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_getters.R
\name{SClusters}
\alias{SClusters}
\title{Gettes for sciClone clusters}
\usage{
SClusters(x, annotations = FALSE)
}
\arguments{
\item{x}{A mvMOBSTER \code{mbst_data} object.}

\item{annotations}{If the annotations available for the mutations should be returned.}
}
\value{
sciClone clustering assigments for every mutation, if available.
}
\description{
Gettes for sciClone clusters
}
\examples{
TODO
}
