% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_print.R
\name{summary.mbst_data}
\alias{summary.mbst_data}
\title{Summary for mvMOBSTER objects.}
\usage{
\method{summary}{mbst_data}(x, ...)
}
\arguments{
\item{x}{A mvMOBSTER \code{mbst_data} object.}

\item{...}{}
}
\value{
Nothing
}
\description{
Summary for mvMOBSTER objects.
}
\examples{
TODO
}
