% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_long_format.R
\name{VAF}
\alias{VAF}
\title{Extract VAF values as they are stored inside an \code{evoverse} dataset.}
\usage{
VAF(x, ids = keys(x), samples = x$samples)
}
\arguments{
\item{x}{An `evoverse` object.}

\item{ids}{The id of the mutations to extract, all by default.}

\item{samples}{The samples for which we extract the data, all by default.}
}
\value{
A tibble of the required entries.
}
\description{
Extract from the internal representation of an object all the entries
that refer to the VAF values. These can be subset by sample and mutation
id; by default all entries are returned. The output is a tibble; no other
transformations are executed.
}
\examples{
data('example_evoverse')

# Default is everything
VAF(example_evoverse)

# Only one sample
VAF(example_evoverse, samples = example_evoverse$samples[1])
}
