% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other_wrappers.R
\name{wrap_neutralitytestr}
\alias{wrap_neutralitytestr}
\title{Run \code{neutralitytestr}}
\usage{
wrap_neutralitytestr(x, samples = x$samples, ql = 0.05, qr = 0.95)
}
\arguments{
\item{x}{A mvMOBSTER \code{mbst_data} object.}

\item{samples}{The samples to use.}
}
\value{
The list of outputs from running \code{neutralitytest},
and a \code{ggpubr} plot.
}
\description{
Wrap a call to \code{neutralitytest} from the \code{neutralitytestr}
package, which measures the slope of the tail in the MOBSTER
fits inside \code{x}, if any. Some evolutionary paraeters are
returned as well. Use the functions available in the \code{MOBSTER}
package to obtain also other types of evolutionary parameters.
}
\examples{
TODO
}
