% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_wide_format.R
\name{Annotations}
\alias{Annotations}
\title{Extract all the annotations available for the mutations stored inside an \code{evoverse} dataset.}
\usage{
Annotations(x, ids = keys(x))
}
\arguments{
\item{x}{An `evoverse` object.}

\item{ids}{The id of the mutations to extract, all by default.}
}
\value{
A tibble of the required entries.
}
\description{
Extract from the internal representation of an object all the entries
that do not refer to any of VAF, DP and NV values, chromsome locations.

This getter differs from the one for DP, NV and VAF values. In particular
since annotations are the same across all samples, this function does not
have samples as parameters. Moreover, in this case the data is stored
in the wide format inside \code{evoverse}, and this functions returns
that type of representation.
}
\examples{
data('example_evoverse')

# Default is everything
Annotations(example_evoverse)

# Only one sample
Annotations(example_evoverse)
}
