% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clusters.R
\name{Clusters}
\alias{Clusters}
\title{Extract clustering results.}
\usage{
Clusters(x)
}
\arguments{
\item{x}{A mvMOBSTER \code{mbst_data} object.}
}
\value{
The input data and all computed clustering assigments for every mutation.
}
\description{
This function extracts the data included into the dataset object by using
function \code{Data_table}, and then attempts extracting clustering from both MOBSTER
and VIBER analyses. If none of these are found, a warning is raised and data are returned.
Otherwise, the data tibble is augmented to have one column for each one of the available
clusters. If a sample is named `X`, column `X.MOBSTER_clusters` reports the
clustering assigments from a MOBSTER analysis of the sample. Instead, VIBER clusters
(multivariate), are reported as a unique column `X_Y_Z.VIBER_clusters` where `X`, `Y`
and `Z` are the sample names of the input dataset.
}
\examples{
TODO
}
