% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_getters.R
\name{Data_table}
\alias{Data_table}
\title{Tabular VAF, DP and NV values, with annotations.}
\usage{
Data_table(x, ids = keys(x), samples = x$samples,
  annotations = FALSE)
}
\arguments{
\item{x}{A mvMOBSTER \code{mbst_data} object.}

\item{ids}{The IDs of the mutations to extract, all by default.}

\item{samples}{The samples for which we extract the data, all by default.}

\item{annotations}{False by default; if true annotations are also returned.}
}
\value{
A spread tibble of the required entries.
}
\description{
This is just a wrapper to a combined call of function \code{VAF_table}, \code{DP_table} and
\code{NV_table}. The resulting outputs are bound by column; ususal subset options are available.
The output can be augmented with annotations from each available mutation.
}
\examples{
TODO
}
