% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_downsample.R
\name{filter_downsample}
\alias{filter_downsample}
\title{Randomly downsample a dataset.}
\usage{
filter_downsample(x, n)
}
\arguments{
\item{x}{An `evoverse` object.}

\item{n}{The number of mutations to retain. If \code{x} contains less
than \code{n} mutations, \code{x} is returned.}
}
\value{
A dataset with at most \code{n} mutations.
}
\description{
Randomly downsample a dataset to contain a certain number of mutations.
If the dataset contains less than the requirednumber of mutations, the input object is returned.
}
\examples{
data('example_evoverse')

N(example_evoverse)

N(filter_downsample(example_evoverse, n = 100))
}
