% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dp_range.R, R/filter_nv_range.R
\name{filter_dp_range}
\alias{filter_dp_range}
\title{Filter all mutations in a DP range}
\usage{
filter_dp_range(x, min.NV, max.NV)

filter_dp_range(x, min.NV, max.NV)
}
\arguments{
\item{x}{A dataset of class \code{"mbst_data"}.}

\item{min.NV}{Minimum NV}

\item{max.NV}{Maximum NV}

\item{min.DP}{Minimum DP}

\item{max.DP}{Maximum DP}

\item{x}{A dataset of class \code{"mbst_data"}.}
}
\value{
A dataset with no mutations outside the
required DP range.

A dataset with no mutations outside the
required NV range.
}
\description{
Filter all mutations that, when they have
DP above 0, they are not in a certain range, given by
\code{min.DP}, \code{max.DP}.

Filter all mutations that, when they have
DP above 0, they are not in a certain range, given by
\code{min.NV}, \code{max.NV}.
}
\examples{
TODO
TODO
}
