% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_getters.R
\name{keys}
\alias{keys}
\title{Get the keys to access mutations.}
\usage{
keys(x)
}
\arguments{
\item{x}{A mvMOBSTER \code{mbst_data} object.}
}
\value{
The keys that id mutations.
}
\description{
Every mutation has its own key which is used to
access and track the mutation. With this function you can
extract the keys. Many functions accept them as parameter to
extract data.
}
\examples{
TODO
}
