% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_getters.R
\name{N}
\alias{N}
\title{Return the number of mutations in the dataset.}
\usage{
N(x)
}
\arguments{
\item{x}{A mvMOBSTER \code{mbst_data} object.}
}
\value{
The number of mutations in the dataset.
}
\description{
Return the number of mutations in the dataset.
}
\examples{
TODO
}
