% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_data.R
\name{NV}
\alias{NV}
\title{Extract the number of reads with the alternative allele from a MOBSTER dataset.}
\usage{
NV(x, ids = keys(x), samples = x$samples)
}
\arguments{
\item{x}{A mvMOBSTER \code{mbst_data} object.}

\item{ids}{The IDs of the mutations to extract, all by default.}

\item{samples}{The samples for which we extract the data, all by default.}
}
\value{
A tibble of the required entries
}
\description{
Extract from the internal representation of an object all the entries
that refer to the NV values. These can be subset by sample and mutation
ID; by default all entries are returbed. The output is a tibble; no other
transformations are executed.
}
\examples{
TODO
}
