% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_data.R
\name{NV_table}
\alias{NV_table}
\title{Tabular NV values.}
\usage{
NV_table(x, ids = keys(x), samples = x$samples, suffix = ".NV")
}
\arguments{
\item{x}{A mvMOBSTER \code{mbst_data} object.}

\item{ids}{The IDs of the mutations to extract, all by default.}

\item{samples}{The samples for which we extract the data, all by default.}

\item{suffix}{Every column will be called as \code{sample_name.NV} when \code{suffix = '.NV'}}
}
\value{
A spread tibble of the required entries.
}
\description{
Similarly to \code{NV}, this function however returns a spread-like table
with one column per VAF value. As \code{NV} the entries can be subset as
required, and the columns named appending a custom suffix to sample names.
}
\examples{
TODO
}
