% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIPELINE_CHR_TIMING.R
\name{pipeline_chromosome_timing}
\alias{pipeline_chromosome_timing}
\title{Pipeline to time aneuploidy.}
\usage{
pipeline_chromosome_timing(
  mutations,
  cna,
  purity,
  karyotypes = c("2:0", "2:1", "2:2"),
  reference = "GRCh38",
  min_muts = 150,
  min_VAF = 0.05,
  description = "Chromosomal timing sample",
  N_max = 15000,
  ...
)
}
\arguments{
\item{mutations}{Input mutations in the format for package \code{CNAqc}.}

\item{cna}{Absolute Copy Number Alterations in the format for package \code{CNAqc}.}

\item{purity}{Tumour purity, in [0, 1].}

\item{karyotypes}{Karyotypes that can be timed, expressed as "Major:minor" notation. Default
are the following states \code{timeable = c('2:0', '2:1', '2:2')}.}

\item{min_muts}{Skip analysing karyotypes with less than \code{min_muts} mutations.}

\item{...}{Parameters passed \code{mobster_fit} in package \code{mobster}.}
}
\value{

}
\description{
For a list of input mutations, CNA and purity data from a tumour sample,
MOBSTER fits are computed to determine proportions of mutations in each
aneuploidy state (copy state). Assignments and plots are returned, and
an automatic QC pipeline (supervised logistic classifier) is used to
flag which fits pass quality check. The trained set for the classifier
consists in n = 377 colorectal whole genome samples.
}
\examples{
# We use data released with the CNAqc package

cna = CNAqc::example_dataset_CNAqc$cna
mutations = CNAqc::example_dataset_CNAqc$snvs
purity = CNAqc::example_dataset_CNAqc$purity

x = pipeline_chromosome_timing(mutations, cna = cna, purity = purity, auto_setup = 'FAST', N_max = 500)
print(x)
}
