% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_qc_copynumbercalls.R
\name{pipeline_qc_copynumbercalls}
\alias{pipeline_qc_copynumbercalls}
\title{Title}
\usage{
pipeline_qc_copynumbercalls(
  mutations,
  cna,
  purity,
  reference = "GRCh38",
  sample = "My Sample",
  matching_epsilon_peaks = 0.025
)
}
\arguments{
\item{matching_epsilon_peaks}{}
}
\value{

}
\description{
Title
}
