% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_qc_copynumbercalls.R
\name{pipeline_qc_copynumbercalls}
\alias{pipeline_qc_copynumbercalls}
\title{Title}
\usage{
pipeline_qc_copynumbercalls(mutations, cna, purity, reference = "GRCh38",
  sample = "PAT00001", output = paste0(sample,
  "_evoverse_qcreport.pdf"), matching_epsilon_peaks = 0.025,
  smooth = FALSE, cex = 0.7, collate = FALSE)
}
\arguments{
\item{collate}{}
}
\description{
Title
}
