% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIPELINE_QC_CALLS.R
\name{pipeline_qc_copynumbercalls}
\alias{pipeline_qc_copynumbercalls}
\title{Pipeline for QC of somatic calls (copy number and mutations).}
\usage{
pipeline_qc_copynumbercalls(
  mutations,
  cna,
  purity,
  reference = "GRCh38",
  description = "My sample",
  smooth = TRUE,
  matching_epsilon_peaks = 0.025,
  ccf_method = "ROUGH",
  peak_method = "closest",
  min_CCF = 0.1,
  only_SNVs = TRUE
)
}
\arguments{
\item{mutations}{Somatic mutation calls, in \code{CNAqc} input format.}

\item{cna}{Absolute Copy Number segments, in \code{CNAqc} input format.}

\item{purity}{Tumour purity, in \code{CNAqc} input format.}

\item{reference}{A reference supported by package \code{CNAqc}.}

\item{description}{A sample description; will appear in the plots.}

\item{smooth}{If \code{TRUE}, smooths copy number data with \code{CNAqc}.}

\item{matching_epsilon_peaks}{Epsilon to match peaks with \code{CNAqc}.}

\item{only_SNVs}{}
}
\value{

}
\description{
This is ...

All input data must be in the format supported by package \code{CNAqc}.
}
