% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_1D_VAF.R
\name{plot_1D_VAF}
\alias{plot_1D_VAF}
\title{Plot a 1-dimensional VAF histogram.}
\usage{
plot_1D_VAF(x, s = x$samples[1])
}
\arguments{
\item{x}{An `evoverse` object.}

\item{s}{The sample name, by default `x$samples[1]` is the first sample in the data.}
}
\value{
A ggplot object.
}
\description{
This function plots a one dimensional
histogam of the input VAF, coloured by mutation karyotype,
for a required input sample.
}
\examples{
data('example_evoverse')

# Plot the 1st (default parameters)
plot_1D_VAF(example_evoverse)

plot_1D_VAF(example_evoverse, s = example_evoverse$samples[2])
}
