% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2D_VAF.R
\name{plot_2D_VAF}
\alias{plot_2D_VAF}
\title{Plot a 2-dimensional VAF scatterplot.}
\usage{
plot_2D_VAF(x, s1 = x$samples[1], s2 = x$samples[2], N = 10000)
}
\arguments{
\item{x}{A `mvMOSTER` object.}

\item{s1}{The first sample name, by default `x$samples[1]` the first sample in the data.}

\item{N}{Maximum number of points to plot, the overall percentage is reported.}

\item{s1}{The second sample name, by default `x$samples[1]` the second sample in the data.}
}
\value{
A `ggplot` object plot.
}
\description{
This function plots a 2D scatterplot
of the input VAF values in sample`s1` versus `s2`.
This function can also subsample the number of plot
points, which helps to allow for a faster rendering
of the returned plot. The number of mutations shown
is always reported, as well as the proportion with
respect to the total number of mutations.
}
\examples{
TODO
}
\seealso{
Function \link{\code{plot_2D_VAF_MOBSTER}} extends this plot visualizing also
MOBSTER clusters' results.
}
