% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIPELINE_QC_CALLS.R
\name{print.evopipe_qc}
\alias{print.evopipe_qc}
\title{S3 print for the pipeline results}
\usage{
\method{print}{evopipe_qc}(x, ...)
}
\arguments{
\item{x}{}
}
\description{
S3 print for the pipeline results
}
