% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_print.R
\name{print.mbst_data}
\alias{print.mbst_data}
\title{Print for mvMOBSTER objects.}
\usage{
\method{print}{mbst_data}(x, ...)
}
\arguments{
\item{x}{A mvMOBSTER \code{mbst_data} object.}

\item{...}{}
}
\value{
Nothing
}
\description{
Print for mvMOBSTER objects.
}
\examples{
TODO
}
