% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_data_size.R
\name{S}
\alias{S}
\title{Return the number of samples in the dataset.}
\usage{
S(x)
}
\arguments{
\item{x}{An `evoverse` object.}
}
\value{
The number of samples in the dataset.
}
\description{
Return the number of samples in the dataset.
}
\examples{
data('example_evoverse')
S(example_evoverse)
}
