% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_print.R
\name{summary.mbst_data}
\alias{summary.mbst_data}
\title{Summary for an \code{evoverse} dataset.}
\usage{
\method{summary}{mbst_data}(x, ...)
}
\arguments{
\item{x}{An  \code{evoverse} object.}

\item{...}{S3 parameters}
}
\value{
Nothing.
}
\description{
The summary reports basic statistics on the stored
input values concerning both mutations, segments, and fits if
available
}
\examples{
data('example_evoverse')
summary(example_evoverse)
}
